Split mantle/codec into files.
